fix: actually fix missing files in package
This was due to a bug with npm https://github.com/npm/cli/issues/6164.

Locally I could not replicate because normally I use pnpm, and pnpm uses
npm-packlist to get the proper files to pack. Unfortunately
semantic-release doesn't support pnpm, so for now the files have been
specified manually. This commit can be reverted when this is fixed or
semantic release supports pnpm.